+++ src/components/theme/View/ListingView.jsx
@@ -7,7 +7,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Helmet from 'react-helmet';
 import { Link } from 'react-router-dom';
-import { Container } from 'semantic-ui-react';
+import { Container, Image } from 'semantic-ui-react';
 
 /**
  * List view component class.
@@ -26,6 +26,14 @@ const ListingView = ({ content }) => (
               {item.title}
             </Link>
           </h2>
+          {item.image && (
+            <Image
+              clearing
+              floated="right"
+              alt={item.image_caption ? item.image_caption : item.title}
+              src={item.image.scales.thumb.download}
+            />
+          )}
           {item.description && <p>{item.description}</p>}
         </article>
       ))}